backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pandas>=2.2.1
openpyxl>=3.1.2
python-dotenv>=1.0.1
